refactor(capitalcall): fix test and update all to capitalcall wording
