+++ api/services/trip/src/actions/ExportCapitalCallsAction.ts
@@ -4,8 +4,8 @@ import { BucketName, S3StorageProvider } from '@pdc/provider-file';
 import { internalOnlyMiddlewares } from '@pdc/provider-middleware';
 import { endOfPreviousMonthDate, startOfPreviousMonthDate } from '../helpers/getDefaultDates';
 import { ResultInterface as Campaign } from '../shared/policy/find.contract';
-import { handlerConfig, ParamsInterface, ResultInterface } from '../shared/trip/excelExport.contract';
-import { alias } from '../shared/trip/excelExport.schema';
+import { handlerConfig, ParamsInterface, ResultInterface } from '../shared/capitalcall/export.contract';
+import { alias } from '../shared/capitalcall/export.schema';
 import { BuildExcel } from './excel/BuildExcel';
 import { CheckCampaign } from './excel/CheckCampaign';
 import { GetCampaignInvolvedOperator } from './excel/GetCampaignInvolvedOperators';
