+++ shared/capitalcall/list.schema.ts
@@ -1,0 +1,11 @@
+export const schema = {
+  type: 'object',
+  additionalProperties: false,
+  properties: {
+    territory_id: { macro: 'serial' },
+    operator_id: { macro: 'serial' },
+  },
+};
+
+export const alias = 'capitalcall.export';
+export const binding = [alias, schema];
